Revert "Throttling views branch"
