+++ car-reviews/src/components/MainPage/maincontent.js
@@ -15,16 +15,9 @@ class MainContent extends Component {
         popularCars: [],
         reviews: [],
         popularReviewers: [],
-        counter: 0,
-        newdate: new Date,
-        olddate: null,
       };
 
     componentDidMount() {
-
-        //conditional rendering based on that date
-        
-        this.getUserCounter();
         const localcarsURL = "http://localhost:3001/api/popular/popular_cars";
         const localreviewsURL = "http://localhost:3001/api/popular/featured_reviews";
         const popularReviewersURL = "http://localhost:3001/api/popular/popular_reviewers"
@@ -45,75 +38,9 @@ class MainContent extends Component {
         });
     }
 
-    updateUserCounter = () => {
-        const counter = this.state.counter;
-
-        const newDate = this.state.newdate;
-        const oldDate = this.state.olddate;
-
-
-        console.log('the new date is', newDate);
-        console.log('the ols date is', oldDate);
-
-        // this.getUserCounter();
-        console.log('the counter is ',counter);
-        const config = {
-          headers: { 'jwt': localStorage.getItem('jwt') }
-        };
-        axios.put('http://localhost:3001/api/users/data', { counter }, config)
-          .then(response => {
-            console.log(response);
-            const newstate = {counter: counter + 1}
-            console.log('new response', newstate)
-            this.setState(newstate);
-            
-            if((this.state.counter > 3 && !response.data.paid) || (oldDate === newDate &&  this.state.counter > 3 && !response.data.paid )) {
-                alert('Please pay for a subscription or come back tommorow for more free reviews!')
-                window.location = '/';
-                // return console.log('to many views');
-            } else if(this.state.counter <= 3 || response.data.paid) {
-                //do nothing
-            }
-
-          })
-          .catch(err => {
-            console.warn(err);
-          });
-
-          
-        }
-
-
-        getUserCounter = () => {
-            axios
-              .get('http://localhost:3001/api/users/data', {
-                headers: {
-                  JWT: localStorage.getItem('jwt')
-                }
-              })
-              .then(response => {
-                
-
-                if(this.state.olddate !== this.state.newdate){ 
-                    console.log('The dates do not match!')
-                    const newstate = {counter: 0, olddate: response.data.date}
-                    this.setState(newstate)
-                } else {
-                    const newstate = {counter: response.data.timesViewed, olddate: response.data.date}
-                    this.setState(newstate)
-                }
-                
-                
-              })
-              .catch(err => console.warn(err));
-          };
-
-    
-
     render() { 
         return (
             <div className="main-content-container" >
-
             <div className="main-content">
             <div style={{ height: '20px'}}></div>
                 <Container>
@@ -121,7 +48,7 @@ class MainContent extends Component {
                     <Row>
                         {this.state.reviews.map(review => {
                             return (
-                                <Col lg="3" md="6" key={review._id} onClick ={this.updateUserCounter}>
+                                <Col lg="3" md="6" key={review._id}>
                                     <ReviewModal {...review}/>
                                 </Col>
                             );
