+++ car-reviews/src/components/Modals/reviewmodal.js
@@ -1,9 +1,7 @@
 import React, { Component } from 'react';
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter, CardText } from 'reactstrap';
-import placeholder from '../../logo.svg';
-// import f150 from '../../f150.jpg';
+import { Button, Modal, ModalHeader, ModalBody, ModalFooter, CardText, Alert } from 'reactstrap';
 import axios from 'axios';
-// import '../MainPage/mainpage.css';
+import ReactStars from 'react-stars'
 import './reviewmodal.css';
 
 // This component is the review modal. It is rendered in maincontent.js
@@ -13,7 +11,8 @@ class ReviewModal extends Component {
     super(props);
     this.state = {
       modal: false,
-      
+      counter: 0,
+      paid: false
     };
 
     this.toggle = this.toggle.bind(this);
@@ -21,17 +20,82 @@ class ReviewModal extends Component {
 
   toggle() {
     // this.modelOpen();
-    this.setState({
-      modal: !this.state.modal
-    });
+    if(this.state.counter > 3 && !this.state.paid) {
+      alert('Please pay for a subscription or come back tommorow for more free reviews!')
+      return console.log('to many views');
+    } else if(this.state.counter <= 3 || this.state.paid) {
+
+      this.modelOpen();
+      this.setState({
+        modal: !this.state.modal
+      });
+    } else {
+      console.log('there was a problem');
+    }
+    
   }
 
+  modelOpen() {
+    if(this.state.modal === true) {
+      this.getUserCounter();
+      this.updateUserCounter();
+    }
+  }
+
+
+  
+
+
+
+  getUserCounter = () => {
+    // const newReview = this.state['review'];
+    // const requestURL = 'https://back-lambda-car-reviews.herokuapp.com/api/reviews';
+    // const localRequests = 'http://localhost:3001/api/reviews';
+    // const counter = this.state.counter;
+    axios
+      .get('http://localhost:3001/api/users/data', {
+        headers: {
+          JWT: localStorage.getItem('jwt')
+        }
+      })
+      .then(response => {
+        // console.log(response);
+        console.log("Times Viewed:", response.data.timesViewed);
+        const newstate = {counter: response.data.timesViewed}
+        this.setState(newstate)
+        
+      })
+      .catch(err => console.warn(err));
+  };
+  
+  updateUserCounter = () => {
+    const counter = this.state.counter;
+    
+    console.log('the counter is ',counter);
+    const config = {
+      headers: { 'jwt': localStorage.getItem('jwt') }
+    };
+    axios.put('http://localhost:3001/api/users/data', { counter }, config)
+      .then(response => {
+        console.log("USER view count", response);
+        const newstate = {counter: counter + 1, paid: response.data.paid}
+        this.setState(newstate)
+        // if (this.state.alerts.password) this.handleAlerts('password');
+        // if (!this.state.alerts.passwordSuccess) this.handleAlerts('passwordSuccess');
+         //localStorage.setItem('jwt', response.data.JWT);
+      })
+      .catch(err => {
+        // if (!this.state.alerts.password) this.handleAlerts('password');
+        // if (this.state.alerts.passwordSuccess) this.handleAlerts('passwordSuccess');
+        console.warn(err);
+      });
+    }
+    ///////////////////////////////////////////////
   
   render() {
     const { score, createOn, title, content, carImage } = this.props;
     console.log("MODAL PROPS: ", this.props)
     const { year, make, model, edition } = this.props.car;
-    // const { username } = this.props.user;
     return (
       <div>
         <Button className="modal-button" onClick={this.toggle}>
@@ -68,7 +132,7 @@ class ReviewModal extends Component {
             {/* <p>Rating: {score} out of 5</p> */}
           </ModalHeader>
           <ModalBody className="modal-body">
-            <img src={carImage} style={{ height: '100%', width: '100%' }} />
+            <img src={carImage} style={{ height: '100%', width: '100%' }} alt=""/>
             <p>{title}</p>
             <p>{content}</p>
           </ModalBody>
