+++ car-reviews/src/components/UserPage/reviewlist.js
@@ -58,8 +58,6 @@ class ReviewList extends Component {
   };
 
 
-  
-
   // handleUpdate = id => {
   //   const config = {
   //     headers: {
